requirements.txt
cryptography==41.0.7
dnspython==2.4.2
scapy~=2.5.0